Add files via upload
This Data Capstone Project deals with 911 emergency calls data taken from the Kaggle website. It deals with Exploratory data analysis on the dataset. The process undertaken in the project are as follows,
data loading
data Cleaning
data manipulation
data visualization

Using the apply and Lambda method the data manipulation is done and the main purpose of the analysis is done with time stream data variables using other categorical variables. The visualization is done by using different libraries like Matplotlib, Seaborn.